+++ base/workflow/R/run.write.configs.R
@@ -1,17 +1,6 @@
-#-------------------------------------------------------------------------------
-# Copyright (c) 2012 University of Illinois, NCSA.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the 
-# University of Illinois/NCSA Open Source License
-# which accompanies this distribution, and is available at
-# http://opensource.ncsa.illinois.edu/license.html
-#-------------------------------------------------------------------------------
-
 ##' Main driver function to call the ecosystem model specific (e.g. ED, SiPNET) 
 ##' run and configuration file scripts 
 ##' 
-##' DEPRECATED: This function has been moved to the PEcAn.workflow package and will be removed from PEcAn.utils.
-##'
 ##' @name run.write.configs
 ##' @title Run model specific write configuration functions
 ##' @param model the ecosystem model to generate the configuration files for
@@ -30,7 +19,6 @@
 run.write.configs <- function(settings, write = TRUE, ens.sample.method = "uniform", 
                               posterior.files = rep(NA, length(settings$pfts)), 
                               overwrite = TRUE) {
-  .Deprecated("PEcAn.workflow::run.write.configs")
   
   con <- PEcAn.DB::db.open(settings$database$bety)
   on.exit(PEcAn.DB::db.close(con))
@@ -151,23 +139,4 @@ run.write.configs <- function(settings, write = TRUE, ens.sample.method = "unifo
   options(scipen = scipen)
   
   return(invisible(settings))
-} # run.write.configs
-
-
-##' @export
-runModule.run.write.configs <- function(settings, overwrite = TRUE) {
-  .Deprecated("PEcAn.workflow::runModule.run.write.configs")
-  if (PEcAn.settings::is.MultiSettings(settings)) {
-    if (overwrite && file.exists(file.path(settings$rundir, "runs.txt"))) {
-      PEcAn.logger::logger.warn("Existing runs.txt file will be removed.")
-      unlink(file.path(settings$rundir, "runs.txt"))
-    }
-    return(PEcAn.settings::papply(settings, runModule.run.write.configs, overwrite = FALSE))
-  } else if (PEcAn.settings::is.Settings(settings)) {
-    write <- settings$database$bety$write
-    ens.sample.method <- settings$ensemble$method
-    return(run.write.configs(settings, write, ens.sample.method, overwrite = overwrite))
-  } else {
-    stop("runModule.run.write.configs only works with Settings or MultiSettings")
-  }
-} # runModule.run.write.configs
+}